Split workflow runtime coverage tests
